# docs/api.yaml
openapi: 3.0.3
info:
  title: OR1ON API
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local server
...
paths:
...
  /tours:
    post:
      tags:
        - POST
      summary: Create a new tour
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Tour name
                  example: "Tour1"
                type:
                  type: string
                  description: Tour type
                  example: "Tour1 description"
                highlights:
                  type: array
                  description: Tour highlights
                  items:
                    type: object
                    description: Highlight ID
                  example: [ highlight1, highlight2 ]
      responses:
        '200':
          description: New tour created
        '404':
          description: Invalid tour name, type or description or highlights.
        '500':
          description: Internal server error
  /tours/{tourID}:
    delete:
      tags:
        - DELETE
      summary: Delete a tour by ID
      parameters:
        - name: tourID
          in: path
          required: true
          schema:
            type: integer
            description: ID of the selected tour
      responses:
        '204':
          description: Tour deleted
        '404':
          description: Tour not found
    put:
      tags:
        - PUT
      summary: Update a tour by ID
      parameters:
        - name: tourID
          in: path
          required: true
          schema:
            type: integer
            description: ID of the selected tour
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Tour name
                  example: "Tour1"
                type:
                  type: string
                  description: Tour type
                  example: "Historical"
                highlights:
                  type: array
                  description: Tour highlights
                  items:
                    type: object
                    description: Highlight ID
                  example: [ highlight1, highlight2 ]
      responses:
        '200':
          description: Feedback updated
        '404':
          description: Highlight or Feedback not found
...
  /routes/{routeID}:
    get:
      tags:
        - GET
      summary: Get a route by ID
      parameters:
        - name: routeID
          in: path
          required: true
          schema:
            type: integer
            description: ID of the selected route
      responses:
        '200':
          description: Route found
        '404':
          description: Route not found.
  /routes/{routeID}/highlights:
    get:
      tags:
        - GET
      summary: Get all highlights from the given Route
      parameters:
        - name: routeID
          in: path
          required: true
          schema:
            type: integer
            description: ID of the selected route
      responses:
        '200':
          description: A list of highlights
        '404':
          description: Route not found
...